+++ packages/Tabs/src/components/Panel/Panel.styles.js
@@ -1,6 +1,13 @@
-import styled from "styled-components";
+import styled, { css } from "styled-components";
 import stylers from "@paprika/stylers";
 
 export const Panel = styled.div`
   margin-top: ${stylers.spacer(2)};
+
+  ${({ isSelected, shouldUnmount }) =>
+    !isSelected && !shouldUnmount
+      ? css`
+          display: none;
+        `
+      : ""}
 `;