+++ packages/Tabs/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+- Added `shouldUnmount` prop on `<Tabs.Panel>` to avoid unmount a single panel when switching between.
+
 ## [0.2.17] - 2020-09-02
 
 ### Added